courier-bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0

